+++ changelog.md
@@ -1,10 +1,12 @@
 # Change Log
 
 ## 1.1.0 - 2020/09/14
+- Added the `exclude_substtings` search argument.
 - Added the `structure` argument.
 - Added the `ignore_note_file_names` search argument.
 - Added the `do_in_constructor` argument.
 - Added some public methods.
+- Changed the `exclude_file_names` search option to be a substring of a base name instead of the whole base name without file extension. 
 - Supported no namespace classes.
 
 ## 1.0.1 - 2020/01/20